Change the return value for ResponseCode::high from u16 to u8
The EDNS response code part only contains 8 bit so the return type fits its
need better. This aligns the type with `ResponseCode::from` and
`Edns::set_rcode_high`.

Closes #1202

+++ crates/proto/src/op/response_code.rs
@@ -136,9 +136,9 @@ impl ResponseCode {
         (u16::from(self) & 0x000F) as u8
     }
 
-    /// returns the high 12 bits for the edns portion of the response code
-    pub fn high(self) -> u16 {
-        (u16::from(self) & 0x0FF0) >> 4
+    /// returns the high 8 bits for the EDNS portion of the response code
+    pub fn high(self) -> u8 {
+        ((u16::from(self) & 0x0FF0) >> 4) as u8
     }
 
     /// Combines the EDNS high and low from the Header to produce the Extended ResponseCode
